Moved the actions fired by a specific composable to the composable that fires them.
