+++ app/src/main/java/com/hallett/taskassistant/corndux/_actions.kt
@@ -17,18 +17,3 @@ object LoadLargerScope : Action
 object LoadSmallerScope : Action
 
 object AddRandomOverdueTask : Action
-
-data class ClickTaskInList(val task: Task) : Action
-
-sealed interface TaskAction : Action
-data class DeleteTask(val task: Task) : TaskAction
-data class DeferTask(val task: Task) : TaskAction
-data class RescheduleTask(val task: Task) : TaskAction
-data class MarkTaskAsComplete(val task: Task) : TaskAction
-data class MarkTaskAsIncomplete(val task: Task) : TaskAction
-
-sealed interface ScopeSelectionAction : Action
-data class ClickNewScope(val newTaskScope: Scope?) : ScopeSelectionAction
-data class ClickNewScopeType(val scopeType: ScopeType) : ScopeSelectionAction
-object CancelScopeSelection : ScopeSelectionAction
-object EnterScopeSelection : ScopeSelectionAction
