+++ app/src/main/java/com/hallett/taskassistant/corndux/performers/TaskActionsPerformer.kt
@@ -7,12 +7,14 @@ import com.hallett.corndux.SideEffect
 import com.hallett.database.ITaskRepository
 import com.hallett.domain.model.TaskStatus
 import com.hallett.scopes.scope_generator.IScopeCalculator
-import com.hallett.taskassistant.corndux.DeferTask
-import com.hallett.taskassistant.corndux.DeleteTask
-import com.hallett.taskassistant.corndux.MarkTaskAsComplete
-import com.hallett.taskassistant.corndux.MarkTaskAsIncomplete
-import com.hallett.taskassistant.corndux.RescheduleTask
-import com.hallett.taskassistant.corndux.TaskAction
+import com.hallett.taskassistant.corndux.UpdateExpandedTask
+import com.hallett.taskassistant.ui.composables.ClickTaskInList
+import com.hallett.taskassistant.ui.composables.DeferTask
+import com.hallett.taskassistant.ui.composables.DeleteTask
+import com.hallett.taskassistant.ui.composables.MarkTaskAsComplete
+import com.hallett.taskassistant.ui.composables.MarkTaskAsIncomplete
+import com.hallett.taskassistant.ui.composables.RescheduleTask
+import com.hallett.taskassistant.ui.composables.TaskAction
 
 class TaskActionsPerformer(
     private val taskRepo: ITaskRepository,
@@ -40,6 +42,7 @@ class TaskActionsPerformer(
                 taskRepo.updateStatus(action.task, TaskStatus.COMPLETE)
             }
             is MarkTaskAsIncomplete -> taskRepo.updateStatus(action.task, TaskStatus.INCOMPLETE)
+            is ClickTaskInList -> dispatchCommit(UpdateExpandedTask(action.task))
         }
     }
 }
