+++ app/src/main/java/com/hallett/taskassistant/overdueTasks/corndux/OverduePerformer.kt
@@ -37,7 +37,6 @@ class OverduePerformer(
                     )
                 )
             )
-            is ClickTaskInList -> dispatchCommit(UpdateExpandedTask(action.task))
             is AddRandomOverdueTask -> {
                 taskRepo.randomTask(ScopeType.values().random(), overdue = true)
             }
